requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]